Remove generation of favicon.ico
see #4885

+++ wcfsetup/install/files/lib/data/style/StyleAction.class.php
@@ -330,13 +330,6 @@ class StyleAction extends AbstractDatabaseObjectAction implements IToggleAction
                         $thumbnail = null;
                     }
 
-                    // Create ICO file.
-                    $ico = @new \PHP_ICO($file->getLocation(), [
-                        [16, 16],
-                        [32, 32],
-                    ]);
-                    $ico->save_ico($style->getAssetPath() . "favicon.ico");
-
                     (new StyleEditor($style))->update([
                         'hasFavicon' => 1,
                     ]);
@@ -347,7 +340,10 @@ class StyleAction extends AbstractDatabaseObjectAction implements IToggleAction
                 foreach ($images as $filename => $length) {
                     \unlink($style->getAssetPath() . $filename);
                 }
-                \unlink($style->getAssetPath() . "favicon.ico");
+                if (\file_exists($style->getAssetPath() . "favicon.ico")) {
+                    \unlink($style->getAssetPath() . "favicon.ico");
+                }
+
                 \unlink($style->getAssetPath() . "manifest.json");
                 \unlink($style->getAssetPath() . "browserconfig.xml");
                 foreach (['png', 'jpg', 'gif'] as $extension) {
